Dispatch events based on delegate target
This is similar to an earlier proposed change, events are dispatched by
the renderer based on the `.Target` of the delegate rather than by
componentId.

I also changed the interface used for dispatching state notifications to
the bare minimum that we need since it's used by both bind and event
handlers.

There is a change in behaviour now wherever a component hooks up a
delegate that's bound to an "AppState" or something else that isn't a
component. In that case the current codegen will not be smart enough to
dispatch the change to the rendering component.